refactored : added BigDecimal

+++ src/main/java/exam/dto/SouvenirDto.java
@@ -2,7 +2,8 @@ package exam.dto;
 
 import lombok.NonNull;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public record SouvenirDto(Long id, @NonNull String name, @NonNull LocalDate date, long price) {
+public record SouvenirDto(Long id, @NonNull String name, @NonNull LocalDate date, BigDecimal price) {
 }
